requirements.txt
flask
flask-cors
bcrypt
groq
yagmail
firebase-admin
apscheduler
python-dotenv
